fix: fix region failover test
